requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
billiard==4.2.0
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.2
dnspython==2.5.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.2
fastapi-cache2==0.2.1
fastapi-users==12.1.3
fastapi-users-db-sqlalchemy==6.0.1
flower==2.0.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
humanize==4.9.0
idna==3.6
kombu==5.3.5
makefun==1.15.2
Mako==1.3.2
MarkupSafe==2.1.5
packaging==23.2
passlib==1.7.4
pendulum==3.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pycparser==2.21
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
PyJWT==2.8.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.7
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==4.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
time-machine==2.13.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
uvicorn==0.27.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
